requirements.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
drf_yasg==1.21.5
Pillow==9.5.0
django-cors-headers==4.0.0
firebase_admin==6.1.0
python-dotenv==1.0.0